Added flashing spin, reverted delete_on_hit, added networking
copied and pasted from discord: "i had to make a few changes to add networking. i also added in a simple flashing animation and doubled the damage. i think you may have removed the "delete on hit" stuff which made poking an enemy with the sword delete it, so that was added back too
great PR though! might have to go back and change offsets to fix the sword sticking out a couple pixels in front of the player (my fault, don't remember why that happened) but otherwise it's working really well"