Theif Cat Basic Implementation (Needs Art) (#269)
* Theif Cat Basic Implementation

Thief cat now can see the player in their vision detection "box" and runs to them. They do not run away but honestly, idk if they should. They are pretty weak to begin with and this may be annoying chasing them down for the "kill all baddies" rooms.

* Made movement more fair, speed now changes

It now will more smartly decide what next direction to walk in to avoid cases where it is looking up and immediately snap down and detect a player. In addition, it now "lunges" by changing its speed once it sees a player.

* sprite update

* Thief Cat Sprite Direction Update

* More Thief Cat Behavior

* Thief Cat Finalized

* Thief Cat in Objects.tsx now

* Remove Thief Cat from Shrine

Co-authored-by: TheRetroDragon <andrew@retro-dragon.com>